+++ src/main/kotlin/com/nobl9/nobl9/nobl9/Nobl9Settings.kt
@@ -10,7 +10,7 @@ import java.nio.file.Paths
 class Nobl9State {
     var executable = ""
     var logLevel = "trace"
-    var logFilePath = "this.log"
+    var logFilePath = ""
 }
 
 const val LANGUAGE_SERVER_NAME: String = "nobl9-language-server"