+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-# Changelog
+<!-- Keep a Changelog guide -> https://keepachangelog.com -->
+
+# Nobl9 Plugin Changelog
 
 All notable changes to this project are documented in the
-[GitHub releases](https://github.com/nobl9/nobl9-vscode/releases).
+[GitHub releases](https://github.com/nobl9/nobl9-intellij-platform-plugin/releases).
+
+## [Unreleased]
+### Added
+- Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)